Adding initial build stuff
