+++ .eslintrc
@@ -1,0 +1,9 @@
+{
+ "env": {
+    "node": true,
+    "mocha": true
+  },
+  "extends": [
+    "airbnb-base"
+  ]
+}
