+++ package.json
@@ -1,0 +1,38 @@
+{
+  "name": "SimpleChain",
+  "version": "0.0.1",
+  "main": "dist/simple-chain.js",
+  "repository": "git@github.com:RupieNetwork/SimpleChain.git",
+  "author": "C0deMunk33",
+  "license": "MIT",
+  "private": true,
+  "scripts": {
+    "precommit": "yarn lint",
+    "lint": "eslint ./src",
+    "dev": "nodemon ./examples/test.js --exec babel-node",
+    "test": "yarn lint && mocha ./src/tests",
+    "build": "yarn lint && babel lib -d dist",
+    "serve": "node dist/index.js"
+  },
+  "dependencies": {
+    "jayson": "^2.0.5",
+    "msgpack-lite": "^0.1.26",
+    "rimraf": "^2.6.2",
+    "shelljs": "^0.8.1",
+    "web3": "^1.0.0-beta.29"
+  },
+  "devDependencies": {
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-register": "^6.26.0",
+    "connect": "^3.6.5",
+    "eslint": "^4.17.0",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.6.1",
+    "morgan": "^1.9.0",
+    "nodemon": "^1.14.12"
+  }
+}
