+++ src/index.js
@@ -1,0 +1,9 @@
+import SimpleChain from './SimpleChain';
+import SimpleChainRPCServer from './simpleChainRPCServer';
+import SimpleChainRPCClient from './simpleChainRPCClient';
+
+export {
+   SimpleChain,
+   SimpleChainRPCServer,
+   SimpleChainRPCClient
+};
