+++ src/simpleChainRPCClient.js
@@ -1,14 +1,11 @@
-var jayson = require('jayson');
- 
+import jayson from 'jayson';
 
-var client = jayson.client.http('http://localhost:2702');
-
-
+const client = jayson.client.http('http://localhost:2702');
 
 async function updateTrackedState(newTrackedState) {
 	var thisStateString = JSON.stringify(newTrackedState);
     const res = await client.request("UpdateTrackedState", [thisStateString], (err,data)=>{
-		
+
 		console.log("data");
 		console.log(data);
 		return data;
@@ -25,29 +22,29 @@ client.request('GetTrackedState',[key], (e,r)=>{
 
 function send1000(runCount){
 	client.request('GetTrackedState',[key], (e,r)=>{
-		
+
 		//console.log("tracked");
 		//console.log(r);
 		var newTrackedState = r.result;
 		var batch = [];
 		for(var bulkTestIdx = 0; bulkTestIdx<100; bulkTestIdx++){
-			newTrackedState.obj.number++;				
+			newTrackedState.obj.number++;
 			//updateTrackedState(newTrackedState);
 			batch.push(client.request("UpdateTrackedState", [JSON.stringify(newTrackedState)]));
-			
+
 		}
-		
+
 		client.request(batch, function(err, responses) {
 			if(err) throw err;
-			
+
 			//console.log('responses', responses.length); // all responses together
 			if(runCount >1){
-				
+
 				send1000(runCount-1);
 				/*
 				setTimeout(()=>{
 				send1000(runCount-1);
-				
+
 				}, 250);
 */
 			}
@@ -61,11 +58,11 @@ function send1000(runCount){
 				}, 250);
 				*/
 			}
-				
+
 		});
 	});
-	
-	
+
+
 }
 
 send1000(11);
@@ -81,7 +78,7 @@ client.request('GetBlock', [1], function(err, response) {
 	  console.log(err);
 	  throw err;
   }
-  console.log("Block size: " + response.result.length); 
+  console.log("Block size: " + response.result.length);
 });
 
 client.request('GetStateChange', [15000], function(err, response) {
@@ -89,8 +86,8 @@ client.request('GetStateChange', [15000], function(err, response) {
 	  console.log(err);
 	  throw err;
   }
-  console.log("state #15000"); 
-  console.log(response.result); 
+  console.log("state #15000");
+  console.log(response.result);
 });
 
 */
