+++ src/tests/SimpleChain.test.js
@@ -1,0 +1,13 @@
+import http from 'http';
+import assert from 'assert';
+
+import { SimpleChain } from '../index';
+
+describe('Running simplechain tests', () => {
+  it('should return 200', done => {
+    http.get('http://127.0.0.1:1337', res => {
+      assert.equal(200, res.statusCode);
+      done();
+    });
+  });
+});
